refactor: mertic 함수, 기타 데이터 로드 모듈화
#46
