+++ code/load_data.py
@@ -2,6 +2,7 @@ import pickle as pickle
 
 import pandas as pd
 import torch
+import yaml
 
 
 class RE_Dataset(torch.utils.data.Dataset):
@@ -54,6 +55,13 @@ def load_data(dataset_dir):
     return dataset
 
 
+# yaml 파일 불러오기
+def load_config(config_file):
+    with open(config_file) as file:
+        config = yaml.safe_load(file)
+    return config
+
+
 def tokenized_dataset(dataset, tokenizer):
     """tokenizer에 따라 sentence를 tokenizing 합니다."""
 
@@ -67,3 +75,72 @@ def tokenized_dataset(dataset, tokenizer):
     tokenized_sentences = tokenizer(concat_entity, list(dataset["sentence"]), return_tensors="pt", padding=True, truncation=True, max_length=256, add_special_tokens=True, return_token_type_ids=True)
 
     return tokenized_sentences
+
+
+def label_to_num(label, cfg):
+    num_label = []
+    with open(cfg["path"]["dict_label_to_num"], "rb") as f:
+        dict_label_to_num = pickle.load(f)
+    for v in label:
+        num_label.append(dict_label_to_num[v])
+
+    return num_label
+
+
+def num_to_label(label, cfg):
+    """
+    숫자로 되어 있던 class를 원본 문자열 라벨로 변환 합니다.
+    """
+    origin_label = []
+    with open(cfg["path"]["dict_num_to_label"], "rb") as f:
+        dict_num_to_label = pickle.load(f)
+    for v in label:
+        origin_label.append(dict_num_to_label[v])
+
+    return origin_label
+
+
+def load_test_dataset(dataset_dir, tokenizer):
+    """
+    test dataset을 불러온 후,
+    tokenizing 합니다.
+    """
+    test_dataset = load_data(dataset_dir)
+    test_label = list(map(int, test_dataset["label"].values))
+    # tokenizing dataset
+    tokenized_test = tokenized_dataset(test_dataset, tokenizer)
+    return test_dataset["id"], tokenized_test, test_label
+
+
+label_keys = [
+    "no_relation",
+    "org:top_members/employees",
+    "org:members",
+    "org:product",
+    "per:title",
+    "org:alternate_names",
+    "per:employee_of",
+    "org:place_of_headquarters",
+    "per:product",
+    "org:number_of_employees/members",
+    "per:children",
+    "per:place_of_residence",
+    "per:alternate_names",
+    "per:other_family",
+    "per:colleagues",
+    "per:origin",
+    "per:siblings",
+    "per:spouse",
+    "org:founded",
+    "org:political/religious_affiliation",
+    "org:member_of",
+    "per:parents",
+    "org:dissolved",
+    "per:schools_attended",
+    "per:date_of_death",
+    "per:date_of_birth",
+    "per:place_of_birth",
+    "per:place_of_death",
+    "org:founded_by",
+    "per:religion",
+]
