+++ code/train.py
@@ -1,7 +1,5 @@
 import argparse
-import os
-import pickle as pickle
-import random  # for random seed
+import random
 from datetime import datetime
 
 import numpy as np
@@ -11,6 +9,8 @@ import sklearn
 import torch
 import yaml
 from load_data import *
+from metrics import *
+from pyprnt import prnt
 from sklearn.metrics import accuracy_score, f1_score, precision_score, recall_score
 from transformers import (
     AutoConfig,
@@ -28,8 +28,6 @@ from transformers import (
 
 import wandb
 
-# for earlystopping, wandb
-
 
 def set_seed(seed: int):
     torch.manual_seed(seed)
@@ -41,88 +39,7 @@ def set_seed(seed: int):
     random.seed(seed)
 
 
-def klue_re_micro_f1(preds, labels):
-    """KLUE-RE micro f1 (except no_relation)"""
-    label_list = [
-        "no_relation",
-        "org:top_members/employees",
-        "org:members",
-        "org:product",
-        "per:title",
-        "org:alternate_names",
-        "per:employee_of",
-        "org:place_of_headquarters",
-        "per:product",
-        "org:number_of_employees/members",
-        "per:children",
-        "per:place_of_residence",
-        "per:alternate_names",
-        "per:other_family",
-        "per:colleagues",
-        "per:origin",
-        "per:siblings",
-        "per:spouse",
-        "org:founded",
-        "org:political/religious_affiliation",
-        "org:member_of",
-        "per:parents",
-        "org:dissolved",
-        "per:schools_attended",
-        "per:date_of_death",
-        "per:date_of_birth",
-        "per:place_of_birth",
-        "per:place_of_death",
-        "org:founded_by",
-        "per:religion",
-    ]
-    no_relation_label_idx = label_list.index("no_relation")
-    label_indices = list(range(len(label_list)))
-    label_indices.remove(no_relation_label_idx)
-    return sklearn.metrics.f1_score(labels, preds, average="micro", labels=label_indices) * 100.0
-
-
-def klue_re_auprc(probs, labels):
-    """KLUE-RE AUPRC (with no_relation)"""
-    labels = np.eye(30)[labels]
-
-    score = np.zeros((30,))
-    for c in range(30):
-        targets_c = labels.take([c], axis=1).ravel()
-        preds_c = probs.take([c], axis=1).ravel()
-        precision, recall, _ = sklearn.metrics.precision_recall_curve(targets_c, preds_c)
-        score[c] = sklearn.metrics.auc(recall, precision)
-    return np.average(score) * 100.0
-
-
-def compute_metrics(pred):
-    """validation을 위한 metrics function"""
-    labels = pred.label_ids
-    preds = pred.predictions.argmax(-1)
-    probs = pred.predictions
-
-    # calculate accuracy using sklearn's function
-    f1 = klue_re_micro_f1(preds, labels)
-    auprc = klue_re_auprc(probs, labels)
-    acc = accuracy_score(labels, preds)  # 리더보드 평가에는 포함되지 않습니다.
-
-    return {
-        "micro f1 score": f1,
-        "auprc": auprc,
-        "accuracy": acc,
-    }
-
-
-def label_to_num(label):
-    num_label = []
-    with open(cfg["path"]["dict_label_to_num"], "rb") as f:
-        dict_label_to_num = pickle.load(f)
-    for v in label:
-        num_label.append(dict_label_to_num[v])
-
-    return num_label
-
-
-def save_preds_to_csv(preds, micro_f1, auprc):
+def save_difference(preds, micro_f1, auprc):
     # valid dataset에 대한 predict값과 실제 라벨값을 비교해서 오답파일 생성하는 함수
     difference = pd.read_csv(cfg["path"]["valid_path"])  # 기존 valid_dataset 불러와서 source열 삭제
     difference = difference.drop(columns=["source"])
@@ -181,12 +98,12 @@ class EarlyStoppingCallback(TrainerCallback):
 
 
 def train():
-    set_seed(seed)  # 랜덤시드 세팅 함수
-
     # load model and tokenizer
     MODEL_NAME = cfg["params"]["MODEL_NAME"]
     tokenizer = AutoTokenizer.from_pretrained(MODEL_NAME)
+
     seed = cfg["params"]["seeds"]
+    set_seed(seed)  # 랜덤시드 세팅 함수
 
     wandb.init(
         config=cfg,
@@ -207,10 +124,10 @@ def train():
 
     # load dataset
     train_dataset = load_data(cfg["path"]["train_path"])
-    dev_dataset = load_data(cfg["path"]["valid_path"])  # validation용 데이터는 따로 만드셔야 합니다.
+    dev_dataset = load_data(cfg["path"]["valid_path"])
 
-    train_label = label_to_num(train_dataset["label"].values)
-    dev_label = label_to_num(dev_dataset["label"].values)
+    train_label = label_to_num(train_dataset["label"].values, cfg)
+    dev_label = label_to_num(dev_dataset["label"].values, cfg)
 
     # tokenizing dataset
     tokenized_train = tokenized_dataset(train_dataset, tokenizer)
@@ -281,7 +198,7 @@ def train():
     # difference.csv 파일 출력하기
     pred = trainer.predict(RE_dev_dataset)
     preds = pred.predictions.argmax(-1)
-    save_preds_to_csv(preds, micro_f1, auprc)
+    save_difference(preds, micro_f1, auprc)
 
     # YAML 파일로 저장
     config_data = {"micro_f1": micro_f1, "auprc": auprc}
@@ -289,13 +206,6 @@ def train():
         yaml.dump(config_data, yaml_file)
 
     wandb.finish()
-
-
-# yaml 파일 불러오기
-def load_config(config_file):
-    with open(config_file) as file:
-        config = yaml.safe_load(file)
-    return config
 
 
 def main():
